# Website/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/auth/forgot-password": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Display the password reset link request page",
                "description": "Show the password reset link request page.",
                "operationId": "c1542051fec30e69a162ce17ac1befb0",
                "responses": {
                    "200": {
                        "description": "Password reset link request page rendered successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send a password reset link to the user's email",
                "description": "Handle an incoming password reset link request.",
                "operationId": "90feb463f5c195b043a573b4236115a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect back with status message"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}